Remove the experimental flag requirement for checkpoint/restore.
