+++ cli/command/container/start.go
@@ -46,9 +46,8 @@ func NewStartCommand(dockerCli *command.DockerCli) *cobra.Command {
 	flags.StringVar(&opts.detachKeys, "detach-keys", "", "Override the key sequence for detaching a container")
 
 	flags.StringVar(&opts.checkpoint, "checkpoint", "", "Restore from this checkpoint")
-	flags.SetAnnotation("checkpoint", "experimental", nil)
 	flags.StringVar(&opts.checkpointDir, "checkpoint-dir", "", "Use a custom checkpoint storage directory")
-	flags.SetAnnotation("checkpoint-dir", "experimental", nil)
+
 	return cmd
 }
