+++ experimental/README.md
@@ -35,7 +35,6 @@ true
 
 Docker service logs command to view logs for a Docker service. This is needed in Swarm mode.
 Option to squash image layers to the base image after successful builds.
-Checkpoint and restore support for Containers.
 Metrics (Prometheus) output for basic container, image, and daemon operations.
 
  * The top-level [docker deploy](../docs/reference/commandline/deploy.md) command. The
@@ -44,7 +43,6 @@ Metrics (Prometheus) output for basic container, image, and daemon operations.
  * [External graphdriver plugins](../docs/extend/plugins_graphdriver.md)
  * [Ipvlan Network Drivers](vlan-networks.md)
  * [Distributed Application Bundles](docker-stacks-and-bundles.md)
- * [Checkpoint & Restore](checkpoint-restore.md)
  * [Docker build with --squash argument](../docs/reference/commandline/build.md##squash-an-images-layers---squash-experimental-only)
 
 ## How to comment on an experimental feature
