+++ integration-cli/docker_cli_help_test.go
@@ -118,12 +118,9 @@ func (s *DockerSuite) TestHelpTextVerify(c *check.C) {
 		cmdsToTest = append(cmdsToTest, "network inspect")
 		cmdsToTest = append(cmdsToTest, "network ls")
 		cmdsToTest = append(cmdsToTest, "network rm")
-
-		if testEnv.ExperimentalDaemon() {
-			cmdsToTest = append(cmdsToTest, "checkpoint create")
-			cmdsToTest = append(cmdsToTest, "checkpoint ls")
-			cmdsToTest = append(cmdsToTest, "checkpoint rm")
-		}
+		cmdsToTest = append(cmdsToTest, "checkpoint create")
+		cmdsToTest = append(cmdsToTest, "checkpoint ls")
+		cmdsToTest = append(cmdsToTest, "checkpoint rm")
 
 		// Divide the list of commands into go routines and  run the func testcommand on the commands in parallel
 		// to save runtime of test
